img_px2px_diff/px2px_diff_pc/requirements.txt
scipy
pyserial
matplotlib
pandas
scikit-learn
numpy